Reverted all changes back to version da092ca8f0af99e8ae5e6edebc8b164b3f35eaa5
