+++ src/components/CreateCustomProviderDialog.tsx
@@ -11,7 +11,6 @@ import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Loader2 } from "lucide-react";
 import { useCustomLanguageModelProvider } from "@/hooks/useCustomLanguageModelProvider";
-import { IpcClient } from "@/ipc/ipc_client";
 
 interface CreateCustomProviderDialogProps {
   isOpen: boolean;
@@ -38,7 +37,7 @@ export function CreateCustomProviderDialog({
     setErrorMessage("");
 
     try {
-      await IpcClient.getInstance().llm.createCustomLanguageModelProvider({
+      await createProvider({
         id: id.trim(),
         name: name.trim(),
         apiBaseUrl: apiBaseUrl.trim(),
