+++ src/components/ProModeSelector.tsx
@@ -67,7 +67,7 @@ export function ProModeSelector() {
               <a
                 className="inline-flex items-center justify-center gap-2 rounded-md border border-primary/30 bg-primary/10 px-3 py-2 text-sm font-medium text-primary shadow-sm transition-colors hover:bg-primary/20 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring"
                 onClick={() => {
-                  IpcClient.getInstance().external.openExternalUrl(
+                  IpcClient.getInstance().openExternalUrl(
                     "https://dyad.sh/pro#ai",
                   );
                 }}
