+++ src/components/chat/MessagesList.tsx
@@ -82,20 +82,20 @@ export const MessagesList = forwardRef<HTMLDivElement, MessagesListProps>(
                             versionId: previousAssistantMessage.commitHash,
                           });
                           const chat =
-                            await IpcClient.getInstance().chat.getChat(
+                            await IpcClient.getInstance().getChat(
                               selectedChatId,
                             );
                           setMessages(chat.messages);
                         }
                       } else {
                         const chat =
-                          await IpcClient.getInstance().chat.getChat(selectedChatId);
+                          await IpcClient.getInstance().getChat(selectedChatId);
                         if (chat.initialCommitHash) {
                           await revertVersion({
                             versionId: chat.initialCommitHash,
                           });
                           try {
-                            await IpcClient.getInstance().chat.deleteMessages(
+                            await IpcClient.getInstance().deleteMessages(
                               selectedChatId,
                             );
                             setMessages([]);
@@ -160,7 +160,7 @@ export const MessagesList = forwardRef<HTMLDivElement, MessagesListProps>(
                         shouldRedo = false;
                       } else {
                         const chat =
-                          await IpcClient.getInstance().chat.getChat(selectedChatId);
+                          await IpcClient.getInstance().getChat(selectedChatId);
                         if (chat.initialCommitHash) {
                           console.debug(
                             "Reverting to initial commit hash",
@@ -219,7 +219,7 @@ export const MessagesList = forwardRef<HTMLDivElement, MessagesListProps>(
               Tired of waiting on AI?
               <a
                 onClick={() =>
-                  IpcClient.getInstance().external.openExternalUrl(
+                  IpcClient.getInstance().openExternalUrl(
                     "https://dyad.sh/pro#ai",
                   )
                 }
