--- src/ipc/client/app_client.ts
@@ -1,151 +1,0 @@
-import type { IpcRenderer } from "electron";
-import type {
-  App,
-  ListAppsResponse,
-  CreateAppParams,
-  CreateAppResult,
-  Version,
-  BranchResult,
-  ImportAppParams,
-  ImportAppResult,
-  RenameBranchParams,
-} from "../ipc_types";
-
-export class AppClient {
-  private ipcRenderer: IpcRenderer;
-
-  constructor(ipcRenderer: IpcRenderer) {
-    this.ipcRenderer = ipcRenderer;
-  }
-
-  // Create a new app with an initial chat
-  public async createApp(params: CreateAppParams): Promise<CreateAppResult> {
-    return this.ipcRenderer.invoke("create-app", params);
-  }
-
-  public async getApp(appId: number): Promise<App> {
-    return this.ipcRenderer.invoke("get-app", appId);
-  }
-
-  // Get all apps
-  public async listApps(): Promise<ListAppsResponse> {
-    return this.ipcRenderer.invoke("list-apps");
-  }
-
-  public async readAppFile(appId: number, filePath: string): Promise<string> {
-    return this.ipcRenderer.invoke("read-app-file", {
-      appId,
-      filePath,
-    });
-  }
-
-  // Edit a file in an app directory
-  public async editAppFile(
-    appId: number,
-    filePath: string,
-    content: string,
-  ): Promise<void> {
-    await this.ipcRenderer.invoke("edit-app-file", {
-      appId,
-      filePath,
-      content,
-    });
-  }
-
-  // List all versions (commits) of an app
-  public async listVersions({ appId }: { appId: number }): Promise<Version[]> {
-    try {
-      const versions = await this.ipcRenderer.invoke("list-versions", {
-        appId,
-      });
-      return versions;
-    } catch (error) {
-      // Re-throw to be caught by higher-level error handling (e.g., react-query)
-      throw error;
-    }
-  }
-
-  // Revert to a specific version
-  public async revertVersion({
-    appId,
-    previousVersionId,
-  }: {
-    appId: number;
-    previousVersionId: string;
-  }): Promise<void> {
-    await this.ipcRenderer.invoke("revert-version", {
-      appId,
-      previousVersionId,
-    });
-  }
-
-  // Checkout a specific version without creating a revert commit
-  public async checkoutVersion({
-    appId,
-    versionId,
-  }: {
-    appId: number;
-    versionId: string;
-  }): Promise<void> {
-    await this.ipcRenderer.invoke("checkout-version", {
-      appId,
-      versionId,
-    });
-  }
-
-  // Get the current branch of an app
-  public async getCurrentBranch(appId: number): Promise<BranchResult> {
-    return this.ipcRenderer.invoke("get-current-branch", {
-      appId,
-    });
-  }
-
-  // Delete an app and all its files
-  public async deleteApp(appId: number): Promise<void> {
-    await this.ipcRenderer.invoke("delete-app", { appId });
-  }
-
-  // Rename an app (update name and path)
-  public async renameApp({
-    appId,
-    appName,
-    appPath,
-  }: {
-    appId: number;
-    appName: string;
-    appPath: string;
-  }): Promise<void> {
-    await this.ipcRenderer.invoke("rename-app", {
-      appId,
-      appName,
-      appPath,
-    });
-  }
-
-  public async selectAppFolder(): Promise<{
-    path: string | null;
-    name: string | null;
-  }> {
-    return this.ipcRenderer.invoke("select-app-folder");
-  }
-
-  public async checkAiRules(params: {
-    path: string;
-  }): Promise<{ exists: boolean }> {
-    return this.ipcRenderer.invoke("check-ai-rules", params);
-  }
-
-  public async importApp(params: ImportAppParams): Promise<ImportAppResult> {
-    return this.ipcRenderer.invoke("import-app", params);
-  }
-
-  async checkAppName(params: {
-    appName: string;
-  }): Promise<{ exists: boolean }> {
-    return this.ipcRenderer.invoke("check-app-name", params);
-  }
-
-  public async renameBranch(params: RenameBranchParams): Promise<void> {
-    await this.ipcRenderer.invoke("rename-branch", params);
-  }
-}
