--- src/ipc/client/chat_client.ts
@@ -1,266 +1,0 @@
-import type { IpcRenderer } from "electron";
-import type {
-  Chat,
-  ChatResponseEnd,
-  Message,
-  TokenCountParams,
-  TokenCountResult,
-  AppOutput,
-  App,
-  ChatLogsData,
-  ApproveProposalResult,
-} from "../ipc_types";
-import type { ChatSummary, ProposalResult } from "@/lib/schemas";
-
-export interface ChatStreamCallbacks {
-  onUpdate: (messages: Message[]) => void;
-  onEnd: (response: ChatResponseEnd) => void;
-  onError: (error: string) => void;
-}
-
-export class ChatClient {
-  private ipcRenderer: IpcRenderer;
-  private chatStreams: Map<number, ChatStreamCallbacks>;
-
-  constructor(ipcRenderer: IpcRenderer) {
-    this.ipcRenderer = ipcRenderer;
-    this.chatStreams = new Map();
-    this.setupListeners();
-  }
-
-  private setupListeners() {
-    this.ipcRenderer.on("chat:response:chunk", (data) => {
-      if (
-        data &&
-        typeof data === "object" &&
-        "chatId" in data &&
-        "messages" in data
-      ) {
-        const { chatId, messages } = data as {
-          chatId: number;
-          messages: Message[];
-        };
-        const callbacks = this.chatStreams.get(chatId);
-        if (callbacks) {
-          callbacks.onUpdate(messages);
-        } else {
-          console.warn(
-            `[IPC] No callbacks found for chat ${chatId}`,
-            this.chatStreams,
-          );
-        }
-      } else {
-        console.error(new Error(`[IPC] Invalid chunk data received: ${data}`));
-      }
-    });
-
-    this.ipcRenderer.on("chat:response:end", (payload) => {
-      const { chatId } = payload as unknown as ChatResponseEnd;
-      const callbacks = this.chatStreams.get(chatId);
-      if (callbacks) {
-        callbacks.onEnd(payload as unknown as ChatResponseEnd);
-        console.debug("chat:response:end");
-        this.chatStreams.delete(chatId);
-      } else {
-        console.error(
-          new Error(
-            `[IPC] No callbacks found for chat ${chatId} on stream end`,
-          ),
-        );
-      }
-    });
-
-    this.ipcRenderer.on("chat:response:error", (error) => {
-      console.debug("chat:response:error");
-      if (typeof error === "string") {
-        for (const [chatId, callbacks] of this.chatStreams.entries()) {
-          callbacks.onError(error);
-          this.chatStreams.delete(chatId);
-        }
-      } else {
-        console.error("[IPC] Invalid error data received:", error);
-      }
-    });
-  }
-
-  public async getChat(chatId: number): Promise<Chat> {
-    try {
-      const data = await this.ipcRenderer.invoke("get-chat", chatId);
-      return data;
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  // Get all chats
-  public async getChats(appId?: number): Promise<ChatSummary[]> {
-    try {
-      const data = await this.ipcRenderer.invoke("get-chats", appId);
-      // Assuming ChatSummariesSchema.parse is handled elsewhere or removed for modularity
-      return data;
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  // New method for streaming responses
-  public streamMessage(
-    prompt: string,
-    options: {
-      chatId: number;
-      redo?: boolean;
-      attachments?: File[];
-      onUpdate: (messages: Message[]) => void;
-      onEnd: (response: ChatResponseEnd) => void;
-      onError: (error: string) => void;
-    },
-  ): void {
-    const { chatId, redo, attachments, onUpdate, onEnd, onError } = options;
-    this.chatStreams.set(chatId, { onUpdate, onEnd, onError });
-
-    // Handle file attachments if provided
-    if (attachments && attachments.length > 0) {
-      // Process each file and convert to base64
-      Promise.all(
-        attachments.map(async (file) => {
-          return new Promise<{ name: string; type: string; data: string }>(
-            (resolve, reject) => {
-              const reader = new FileReader();
-              reader.onload = () => {
-                resolve({
-                  name: file.name,
-                  type: file.type,
-                  data: reader.result as string,
-                });
-              };
-              reader.onerror = () =>
-                reject(new Error(`Failed to read file: ${file.name}`));
-              reader.readAsDataURL(file);
-            },
-          );
-        }),
-      )
-        .then((fileDataArray) => {
-          // Use invoke to start the stream and pass the chatId and attachments
-          this.ipcRenderer
-            .invoke("chat:stream", {
-              prompt,
-              chatId,
-              redo,
-              attachments: fileDataArray,
-            })
-            .catch((err) => {
-              onError(String(err));
-              this.chatStreams.delete(chatId);
-            });
-        })
-        .catch((err) => {
-          onError(String(err));
-          this.chatStreams.delete(chatId);
-        });
-    } else {
-      // No attachments, proceed normally
-      this.ipcRenderer
-        .invoke("chat:stream", {
-          prompt,
-          chatId,
-          redo,
-        })
-        .catch((err) => {
-          onError(String(err));
-          this.chatStreams.delete(chatId);
-        });
-    }
-  }
-
-  // Method to cancel an ongoing stream
-  public cancelChatStream(chatId: number): void {
-    this.ipcRenderer.invoke("chat:cancel", chatId);
-    const callbacks = this.chatStreams.get(chatId);
-    if (callbacks) {
-      this.chatStreams.delete(chatId);
-    } else {
-      console.error("Tried canceling chat that doesn't exist");
-    }
-  }
-
-  // Create a new chat for an app
-  public async createChat(appId: number): Promise<number> {
-    return this.ipcRenderer.invoke("create-chat", appId);
-  }
-
-  public async deleteChat(chatId: number): Promise<void> {
-    await this.ipcRenderer.invoke("delete-chat", chatId);
-  }
-
-  public async deleteMessages(chatId: number): Promise<void> {
-    await this.ipcRenderer.invoke("delete-messages", chatId);
-  }
-
-  public async addDependency({
-    chatId,
-    packages,
-  }: {
-    chatId: number;
-    packages: string[];
-  }): Promise<void> {
-    await this.ipcRenderer.invoke("chat:add-dep", {
-      chatId,
-      packages,
-    });
-  }
-
-  // Get proposal details
-  public async getProposal(chatId: number): Promise<ProposalResult | null> {
-    try {
-      const data = await this.ipcRenderer.invoke("get-proposal", { chatId });
-      return data as ProposalResult | null;
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  // --- Proposal Management ---
-  public async approveProposal({
-    chatId,
-    messageId,
-  }: {
-    chatId: number;
-    messageId: number;
-  }): Promise<ApproveProposalResult> {
-    return this.ipcRenderer.invoke("approve-proposal", {
-      chatId,
-      messageId,
-    });
-  }
-
-  public async rejectProposal({
-    chatId,
-    messageId,
-  }: {
-    chatId: number;
-    messageId: number;
-  }): Promise<void> {
-    await this.ipcRenderer.invoke("reject-proposal", {
-      chatId,
-      messageId,
-    });
-  }
-  // --- End Proposal Management ---
-
-  public async getChatLogs(chatId: number): Promise<ChatLogsData> {
-    return this.ipcRenderer.invoke("get-chat-logs", chatId);
-  }
-
-  // Count tokens for a chat and input
-  public async countTokens(
-    params: TokenCountParams,
-  ): Promise<TokenCountResult> {
-    try {
-      const result = await this.ipcRenderer.invoke("chat:count-tokens", params);
-      return result as TokenCountResult;
-    } catch (error) {
-      throw error;
-    }
-  }
-}
