--- src/ipc/client/external_integrations_client.ts
@@ -1,157 +1,0 @@
-import type { IpcRenderer } from "electron";
-import type {
-  GitHubDeviceFlowUpdateData,
-  GitHubDeviceFlowSuccessData,
-  GitHubDeviceFlowErrorData,
-  VercelProject,
-  VercelDeployParams,
-} from "../ipc_types";
-
-export class ExternalIntegrationsClient {
-  private ipcRenderer: IpcRenderer;
-
-  constructor(ipcRenderer: IpcRenderer) {
-    this.ipcRenderer = ipcRenderer;
-  }
-
-  // Open an external URL using the default browser
-  public async openExternalUrl(url: string): Promise<void> {
-    await this.ipcRenderer.invoke("open-external-url", url);
-  }
-
-  public async showItemInFolder(fullPath: string): Promise<void> {
-    await this.ipcRenderer.invoke("show-item-in-folder", fullPath);
-  }
-
-  public async uploadToSignedUrl(
-    url: string,
-    contentType: string,
-    data: any,
-  ): Promise<void> {
-    await this.ipcRenderer.invoke("upload-to-signed-url", {
-      url,
-      contentType,
-      data,
-    });
-  }
-
-  // --- GitHub Device Flow ---
-  public startGithubDeviceFlow(appId: number | null): void {
-    this.ipcRenderer.invoke("github:start-flow", { appId });
-  }
-
-  public onGithubDeviceFlowUpdate(
-    callback: (data: GitHubDeviceFlowUpdateData) => void,
-  ): () => void {
-    const listener = (data: any) => {
-      console.log("github:flow-update", data);
-      callback(data as GitHubDeviceFlowUpdateData);
-    };
-    this.ipcRenderer.on("github:flow-update", listener);
-    // Return a function to remove the listener
-    return () => {
-      this.ipcRenderer.removeListener("github:flow-update", listener);
-    };
-  }
-
-  public onGithubDeviceFlowSuccess(
-    callback: (data: GitHubDeviceFlowSuccessData) => void,
-  ): () => void {
-    const listener = (data: any) => {
-      console.log("github:flow-success", data);
-      callback(data as GitHubDeviceFlowSuccessData);
-    };
-    this.ipcRenderer.on("github:flow-success", listener);
-    return () => {
-      this.ipcRenderer.removeListener("github:flow-success", listener);
-    };
-  }
-
-  public onGithubDeviceFlowError(
-    callback: (data: GitHubDeviceFlowErrorData) => void,
-  ): () => void {
-    const listener = (data: any) => {
-      console.log("github:flow-error", data);
-      callback(data as GitHubDeviceFlowErrorData);
-    };
-    this.ipcRenderer.on("github:flow-error", listener);
-    return () => {
-      this.ipcRenderer.removeListener("github:flow-error", listener);
-    };
-  }
-  // --- End GitHub Device Flow ---
-
-  // --- GitHub Repo Management ---
-  public async checkGithubRepoAvailable(
-    org: string,
-    repo: string,
-  ): Promise<{ available: boolean; error?: string }> {
-    return this.ipcRenderer.invoke("github:is-repo-available", {
-      org,
-      repo,
-    });
-  }
-
-  public async createGithubRepo(
-    org: string,
-    repo: string,
-    appId: number,
-  ): Promise<void> {
-    await this.ipcRenderer.invoke("github:create-repo", {
-      org,
-      repo,
-      appId,
-    });
-  }
-
-  // Sync (push) local repo to GitHub
-  public async syncGithubRepo(
-    appId: number,
-  ): Promise<{ success: boolean; error?: string }> {
-    try {
-      const result = await this.ipcRenderer.invoke("github:push", { appId });
-      return result as { success: boolean; error?: string };
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  public async disconnectGithubRepo(appId: number): Promise<void> {
-    await this.ipcRenderer.invoke("github:disconnect", {
-      appId,
-    });
-  }
-  // --- End GitHub Repo Management ---
-
-  // --- Supabase Management ---
-  public async listSupabaseProjects(): Promise<any[]> {
-    return this.ipcRenderer.invoke("supabase:list-projects");
-  }
-
-  public async setSupabaseAppProject(
-    project: string,
-    app: number,
-  ): Promise<void> {
-    await this.ipcRenderer.invoke("supabase:set-app-project", {
-      project,
-      app,
-    });
-  }
-
-  public async unsetSupabaseAppProject(app: number): Promise<void> {
-    await this.ipcRenderer.invoke("supabase:unset-app-project", {
-      app,
-    });
-  }
-  // --- End Supabase Management ---
-
-  // --- Vercel Management ---
-  public async listVercelProjects(): Promise<VercelProject[]> {
-    return this.ipcRenderer.invoke("vercel:list-projects");
-  }
-
-  public async deployVercelProject(params: VercelDeployParams): Promise<void> {
-    return this.ipcRenderer.invoke("vercel:deploy-project", params);
-  }
-  // --- End Vercel Management ---
-}
