--- src/ipc/client/llm_client.ts
@@ -1,83 +1,0 @@
-import type { IpcRenderer } from "electron";
-import type {
-  LanguageModelProvider,
-  LanguageModel,
-  CreateCustomLanguageModelProviderParams,
-  CreateCustomLanguageModelParams,
-  LocalModel,
-} from "../ipc_types";
-
-interface DeleteCustomModelParams {
-  providerId: string;
-  modelApiName: string;
-}
-
-export class LlmClient {
-  private ipcRenderer: IpcRenderer;
-
-  constructor(ipcRenderer: IpcRenderer) {
-    this.ipcRenderer = ipcRenderer;
-  }
-
-  public async getLanguageModelProviders(): Promise<LanguageModelProvider[]> {
-    return this.ipcRenderer.invoke("get-language-model-providers");
-  }
-
-  public async getLanguageModels(params: {
-    providerId: string;
-  }): Promise<LanguageModel[]> {
-    return this.ipcRenderer.invoke("get-language-models", params);
-  }
-
-  public async getLanguageModelsByProviders(): Promise<
-    Record<string, LanguageModel[]>
-  > {
-    return this.ipcRenderer.invoke("get-language-models-by-providers");
-  }
-
-  public async createCustomLanguageModelProvider({
-    id,
-    name,
-    apiBaseUrl,
-    envVarName,
-  }: CreateCustomLanguageModelProviderParams): Promise<LanguageModelProvider> {
-    return this.ipcRenderer.invoke("create-custom-language-model-provider", {
-      id,
-      name,
-      apiBaseUrl,
-      envVarName,
-    });
-  }
-
-  public async createCustomLanguageModel(
-    params: CreateCustomLanguageModelParams,
-  ): Promise<void> {
-    await this.ipcRenderer.invoke("create-custom-language-model", params);
-  }
-
-  public async deleteCustomLanguageModel(modelId: string): Promise<void> {
-    return this.ipcRenderer.invoke("delete-custom-language-model", modelId);
-  }
-
-  async deleteCustomModel(params: DeleteCustomModelParams): Promise<void> {
-    return this.ipcRenderer.invoke("delete-custom-model", params);
-  }
-
-  async deleteCustomLanguageModelProvider(providerId: string): Promise<void> {
-    return this.ipcRenderer.invoke("delete-custom-language-model-provider", {
-      providerId,
-    });
-  }
-
-  public async listLocalOllamaModels(): Promise<LocalModel[]> {
-    const response = await this.ipcRenderer.invoke("local-models:list-ollama");
-    return response?.models || [];
-  }
-
-  public async listLocalLMStudioModels(): Promise<LocalModel[]> {
-    const response = await this.ipcRenderer.invoke(
-      "local-models:list-lmstudio",
-    );
-    return response?.models || [];
-  }
-}
