--- src/ipc/client/settings_client.ts
@@ -1,50 +1,0 @@
-import type { IpcRenderer } from "electron";
-import type { UserSettings } from "../../lib/schemas";
-
-export class SettingsClient {
-  private ipcRenderer: IpcRenderer;
-
-  constructor(ipcRenderer: IpcRenderer) {
-    this.ipcRenderer = ipcRenderer;
-  }
-
-  // Get user settings
-  public async getUserSettings(): Promise<UserSettings> {
-    try {
-      const settings = await this.ipcRenderer.invoke("get-user-settings");
-      return settings;
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  // Update user settings
-  public async setUserSettings(
-    settings: Partial<UserSettings>,
-  ): Promise<UserSettings> {
-    try {
-      const updatedSettings = await this.ipcRenderer.invoke(
-        "set-user-settings",
-        settings,
-      );
-      return updatedSettings;
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  // Get allow-listed environment variables
-  public async getEnvVars(): Promise<Record<string, string | undefined>> {
-    try {
-      const envVars = await this.ipcRenderer.invoke("get-env-vars");
-      return envVars as Record<string, string | undefined>;
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  // Reset all - removes all app files, settings, and drops the database
-  public async resetAll(): Promise<void> {
-    await this.ipcRenderer.invoke("reset-all");
-  }
-}
