--- src/ipc/client/system_client.ts
@@ -1,133 +1,0 @@
-import type { IpcRenderer } from "electron";
-import type {
-  NodeSystemInfo,
-  SystemDebugInfo,
-  DoesReleaseNoteExistParams,
-  AppOutput,
-} from "../ipc_types";
-
-export interface AppStreamCallbacks {
-  onOutput: (output: AppOutput) => void;
-}
-
-export class SystemClient {
-  private ipcRenderer: IpcRenderer;
-  private appStreams: Map<number, AppStreamCallbacks>;
-
-  constructor(ipcRenderer: IpcRenderer) {
-    this.ipcRenderer = ipcRenderer;
-    this.appStreams = new Map();
-    this.setupListeners();
-  }
-
-  private setupListeners() {
-    this.ipcRenderer.on("app:output", (data) => {
-      if (
-        data &&
-        typeof data === "object" &&
-        "type" in data &&
-        "message" in data &&
-        "appId" in data
-      ) {
-        const { type, message, appId } = data as unknown as AppOutput;
-        const callbacks = this.appStreams.get(appId);
-        if (callbacks) {
-          callbacks.onOutput({ type, message, appId, timestamp: Date.now() });
-        }
-      } else {
-        console.error(new Error(`[IPC] Invalid app output data received: ${data}`));
-      }
-    });
-  }
-
-  public async reloadEnvPath(): Promise<void> {
-    await this.ipcRenderer.invoke("reload-env-path");
-  }
-
-  // Run an app
-  public async runApp(
-    appId: number,
-    onOutput: (output: AppOutput) => void,
-  ): Promise<void> {
-    await this.ipcRenderer.invoke("run-app", { appId });
-    this.appStreams.set(appId, { onOutput });
-  }
-
-  // Stop a running app
-  public async stopApp(appId: number): Promise<void> {
-    await this.ipcRenderer.invoke("stop-app", { appId });
-  }
-
-  // Restart a running app
-  public async restartApp(
-    appId: number,
-    onOutput: (output: AppOutput) => void,
-    removeNodeModules?: boolean,
-  ): Promise<{ success: boolean }> {
-    try {
-      const result = await this.ipcRenderer.invoke("restart-app", {
-        appId,
-        removeNodeModules,
-      });
-      this.appStreams.set(appId, { onOutput });
-      return result;
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  // Check Node.js and npm status
-  public async getNodejsStatus(): Promise<NodeSystemInfo> {
-    return this.ipcRenderer.invoke("nodejs-status");
-  }
-
-  // Get the main app version
-  public async getAppVersion(): Promise<string> {
-    const result = await this.ipcRenderer.invoke("get-app-version");
-    return result.version as string;
-  }
-
-  public async getSystemDebugInfo(): Promise<SystemDebugInfo> {
-    return this.ipcRenderer.invoke("get-system-debug-info");
-  }
-
-  // Window control methods
-  public async minimizeWindow(): Promise<void> {
-    try {
-      await this.ipcRenderer.invoke("window:minimize");
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  public async maximizeWindow(): Promise<void> {
-    try {
-      await this.ipcRenderer.invoke("window:maximize");
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  public async closeWindow(): Promise<void> {
-    try {
-      await this.ipcRenderer.invoke("window:close");
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  // Get system platform (win32, darwin, linux)
-  public async getSystemPlatform(): Promise<string> {
-    return this.ipcRenderer.invoke("get-system-platform");
-  }
-
-  public async doesReleaseNoteExist(
-    params: DoesReleaseNoteExistParams,
-  ): Promise<{ exists: boolean; url?: string }> {
-    return this.ipcRenderer.invoke("does-release-note-exist", params);
-  }
-
-  async clearSessionData(): Promise<void> {
-    return this.ipcRenderer.invoke("clear-session-data");
-  }
-}
